Big refactor
In attempting to fix the threading issues in the former implementation, it
became apparent that the implementation had some deep structural issues. This
change attempts to address those issues by reassigning responsibilities to the
appropriate classes, and making appropriate use of thread safety primitives,
as well as non-mutable data where appropriate.